+++ route/route_test.go
@@ -13,16 +13,15 @@ import (
 	"time"
 
 	"github.com/facebookgo/inject"
-	"github.com/jirs5/tracing-proxy/collect"
-	"github.com/jirs5/tracing-proxy/config"
-	"github.com/jirs5/tracing-proxy/logger"
-	"github.com/jirs5/tracing-proxy/metrics"
-	"github.com/jirs5/tracing-proxy/transmit"
+	"github.com/opsramp/tracing-proxy/collect"
+	"github.com/opsramp/tracing-proxy/config"
+	"github.com/opsramp/tracing-proxy/logger"
+	"github.com/opsramp/tracing-proxy/metrics"
+	"github.com/opsramp/tracing-proxy/transmit"
 
 	"github.com/gorilla/mux"
-	"github.com/jirs5/tracing-proxy/sharder"
 	"github.com/klauspost/compress/zstd"
-	"github.com/vmihailenco/msgpack/v4"
+	"github.com/opsramp/tracing-proxy/sharder"
 	"google.golang.org/grpc/metadata"
 )