+++ sample/trace_key.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 	"strconv"
 
-	"github.com/jirs5/tracing-proxy/types"
+	"github.com/opsramp/tracing-proxy/types"
 )
 
 type traceKey struct {