+++ sharder/deterministic.go
@@ -10,9 +10,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/jirs5/tracing-proxy/config"
-	"github.com/jirs5/tracing-proxy/internal/peer"
-	"github.com/jirs5/tracing-proxy/logger"
+	"github.com/opsramp/tracing-proxy/config"
+	"github.com/opsramp/tracing-proxy/internal/peer"
+	"github.com/opsramp/tracing-proxy/logger"
 	"github.com/pkg/errors"
 )