+++ sharder/deterministic_test.go
@@ -6,9 +6,9 @@ package sharder
 import (
 	"testing"
 
-	"github.com/jirs5/tracing-proxy/config"
-	"github.com/jirs5/tracing-proxy/internal/peer"
-	"github.com/jirs5/tracing-proxy/logger"
+	"github.com/opsramp/tracing-proxy/config"
+	"github.com/opsramp/tracing-proxy/internal/peer"
+	"github.com/opsramp/tracing-proxy/logger"
 	"github.com/stretchr/testify/assert"
 )